Crate limn_text_layout

Text layout logic.

Reexports

pub use types::Align;

Modules

cursor

glyph

line

types

Structs

Lines
An iterator yielding each line within the given text as a new &str, where the start and end indices into each line are provided by the given iterator.

Enums

Wrap
The way in which text should wrap around the width.

Functions

get_line_rects

get_positioned_glyphs

get_text_height

get_text_size

lines
Produce an iterator yielding each line within the given text as a new &str, where the start and end indices into each line are provided by the given iterator.

pt_to_px
Converts the given font size in "points" to its font size in pixels. assumes 96 dpi display. 1 pt = 1/72"

pt_to_scale
Converts the given font size in "points" to a uniform rusttype::Scale.

px_to_pt

Type Definitions

Font

PositionedGlyph
The RustType PositionedGlyph type.