Function limn_text_layout::cursor::xys_per_line

Important traits for XysPerLine<'a, I>
impl<'a, I> Iterator for XysPerLine<'a, I> where
I: Iterator<Item = (LineInfo, Rect)>, type Item = (Xs<'a, 'a>, Range);

pub fn xys_per_line<'a, I>(
    lines_with_rects: I, 
    font: &'a Font, 
    text: &'a str, 
    font_size: f32
) -> XysPerLine<'a, I>

Every possible cursor position within each line of text yielded by the given iterator.

Yields (xs, y_range), where y_range is the Range occupied by the line across the y axis and xs is every possible cursor position along the x axis