Module limn_text_layout::glyph

Structs

GlyphRects
An iterator yielding the Rect for each char's Glyph in the given text.

GlyphRectsPerLine
An iterator that, for every (line, line_rect) pair yielded by the given iterator, produces an iterator that yields a Rect for every character in that line.

SelectedGlyphRects
Yields a Rect for each selected character in a single line of text.

SelectedGlyphRectsPerLine
Yields an iterator yielding Rects for each selected character in each line of text within the given iterator yielding char Rects.

Functions

index_after_cursor
Find the index of the character that directly follows the cursor at the given cursor_idx.