Module limn_text_layout::line

Structs

Break

LineInfo
Information about a single line of text within a &str.

LineInfos
An iterator yielding an Info struct for each line in the given text wrapped by the given next_break_fn.

LineRects
An iterator yielding a Rect for each line in

SelectedLineRects
An iterator yielding a Rect for each selected line in a block of text.

Functions

width
Produce the width of the given line of text including spaces (i.e. ' ').